+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Updated mysql-lib v5.5.0
 
 ### Changed
+- get_db_tbl: Removed check for pre-MySQL v8.0 when setting dict_key variable, removed print warnings for no databases to process.
 - run_program: Added check for override option (-i) for non-slave database comparsion.
 - data_out: Changed default email subject line.
 - Documentation changes.